requirements.txt
fastapi
uvicorn
pydantic
google-generativeai>=0.7.2
